requirements.txt
requests
python-dotenv
requests-toolbelt
pycryptodome
aiohttp>=3.8.0
